Reset fixture moved to module-level conftest.py.
